pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tinirag"
version = "0.1.0"
description = "Privacy-first RAG CLI — local LLM + SearXNG, zero cloud"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "httpx>=0.27",
    "openai>=1.30",
    "trafilatura>=1.9",
    "markdownify>=0.13",
    "tiktoken>=0.7",
    "python-dotenv>=1.0",
    "tomli>=2.0",
    "tomli_w>=1.0",
]

[project.scripts]
tinirag = "tinirag.cli:app"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "ruff>=0.4",
]
